backend/requirements_clean.txt
Flask==3.0.3
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
openai==1.61.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
pytest==8.3.4
pytest-cov==6.0.0